hosts device

+++ firewall/fw_hosts.h
@@ -3,7 +3,7 @@
 
 #define DEVICE_NAME_HOSTS "hosts"
 
-__u8 check_hosts(char *host);
+int check_hosts(char *host);
 
 int init_hosts(void);
 void cleanup_hosts(void);
